requirements.txt
Flask==3.0.0; python_version > '3.6'
google-cloud-datastore
gunicorn
mechanize
python-dotenv
beautifulsoup4
pandas